Renaming android module.
