+++ app/build.gradle.kts
@@ -1,6 +1,5 @@
 plugins {
     id("com.android.application")
-    id("com.google.dagger.hilt.android")
     id("kotlin-android")
     id("kotlin-kapt")
     id("org.jmailen.kotlinter")
@@ -10,7 +9,7 @@ android {
     compileSdk = libs.versions.compileSdk.get().toInt()
 
     defaultConfig {
-        applicationId = "template.app.id"
+        applicationId = "com.adammcneilly.recipe"
         minSdk = libs.versions.minSdk.get().toInt()
         targetSdk = libs.versions.compileSdk.get().toInt()
         versionCode = 1
@@ -52,7 +51,7 @@ android {
         }
     }
 
-    namespace = "template"
+    namespace = "com.adammcneilly.recipe"
 }
 
 dependencies {
@@ -67,31 +66,20 @@ dependencies {
     implementation(libs.androidx.appcompat)
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
-    implementation(libs.androidx.room.runtime)
     implementation(libs.android.material)
     implementation(libs.compose.ui)
     implementation(libs.compose.material)
     implementation(libs.compose.ui.tooling)
-    implementation(libs.hilt.android)
-    implementation(libs.square.moshi.kotlin)
-    implementation(libs.square.retrofit)
-    implementation(libs.square.retrofit.converter.moshi)
 
     testImplementation(libs.junit)
 
     androidTestImplementation(libs.androidx.test.junit)
     androidTestImplementation(libs.androidx.test.espresso.core)
     androidTestImplementation(libs.compose.ui.test.junit)
-    androidTestImplementation(libs.hilt.android.testing)
 
     debugImplementation(libs.compose.ui.test.manifest)
     debugImplementation(libs.compose.ui.tooling)
     debugImplementation(libs.square.leakcanary)
 
-    kapt(libs.androidx.room.compiler)
-    kapt(libs.hilt.compiler)
-    kapt(libs.square.moshi.kotlin.codegen)
-    kaptAndroidTest(libs.hilt.android.compiler)
 
-    annotationProcessor(libs.androidx.room.compiler)
 }
