+++ app/src/main/java/com/adammcneilly/recipe/theme/Color.kt
@@ -1,6 +1,6 @@
 @file:Suppress("MagicNumber")
 
-package template.theme
+package com.adammcneilly.recipe.theme
 
 import androidx.compose.ui.graphics.Color
