+++ app/src/main/java/com/adammcneilly/recipe/theme/Type.kt
@@ -1,4 +1,4 @@
-package template.theme
+package com.adammcneilly.recipe.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
